Apply expressions to data.frames

Creates a function that evaluates expressions in a future data.frame. Is like with(), but the data argument is passed at a later step.

delayed_with(...)

Arguments

...

Expressions that will be evaluated.

Value

A function that takes a data.frame and returns the expressions in ...

evaluated in an environment constructed from it.

Details

Each expression in ... must return a single numeric value. They can be named or return named vectors.

See also

Examples

some_stats <- delayed_with(mean_x = mean(x), mean(y), sd(x), coef(lm(x ~ y)))
data <- data.frame(x = rnorm(20) , y = rnorm(20))
some_stats(data)
#>      mean_x                         (Intercept)           y 
#> -0.15060314  0.06734492  1.13666509 -0.12370109 -0.39946657