Increase resolution of data

Interpolates between the output of draw_data() and increases the point density of each stroke.Useful for avoiding sparse targets that result in clumping of points when metamerizing. It only has an effect on strokes (made by double clicking).

densify(data, res = 2)

Arguments

data

A data.frame with columns x, y and .group.

res

A numeric indicating the multiplicative resolution (i.e. 2 = double resolution).

Value

A data.frame with the x and y values of your data and a .group column that identifies each stroke.

See also