Mean distance to an sf object

Mean distance to an sf object

mean_dist_to_sf(target, coords = c("x", "y"), buffer = 0, squared = TRUE)

Arguments

target

An sf object.

coords

Character vector with the columns of the data object that define de coordinates.

buffer

Buffer around the sf object. Distances smaller than buffer are replaced with 0.

squared

Logical indicating whether to compute the mean squared distance (if TRUE) or the mean distance.

See also